Clean up deprecations, migrate MIDI API, universal macOS build
Dead code:
- Remove Source/FilterVue copie.h, Source/labo.h, Source/WidgetsDemo.h.pdf

Deprecation warnings (now builds warning-free):
- forEachXmlChildElement -> getChildIterator() (AWMVue, Controller, MidiDemo)
- Font(...) -> FontOptions(...)
- XmlElement::writeToFile -> writeTo
- Font::getStringWidth -> GlyphArrangement::getStringWidth
- Fix long->int precision warning (MidiSysex.h)

MIDI API migration (index-based -> identifier-based):
- MidiDeviceListEntry now stores MidiDeviceInfo (name + identifier)
- getAvailableDevices() / openDevice(identifier) instead of the deprecated
  index-based getDevices() / openDevice(int). Robust to port reordering.

Distribution:
- Universal macOS build (arm64 + x86_64) via customXcodeFlags in Sysex77.jucer
- Ad-hoc codesign the bundle
- Refresh dist/Sysex77.app (universal, signed)
- Update CHANGELOG.md

+++ Source/MidiSysex.h
@@ -219,7 +219,7 @@ void sendRaw(const void* sysexData, const long dataSize)
         Logger::writeToLog(String(dataSize));
         //    Logger::writeToLog(String(dataSize));
         //Envoi midi brut depuis la mémoire
-        MidiMessage m = MidiMessage (sysexData, dataSize);
+        MidiMessage m = MidiMessage (sysexData, (int) dataSize);
         m.setTimeStamp (Time::getMillisecondCounterHiRes() * 0.001);
         sendToOutputs (m);
     }
